+++ src/ui/dialogs/parameterdialog.h
@@ -19,6 +19,8 @@
 
 #pragma once
 
+#include <charts.h>
+
 #include "src/global.h"
 
 #include <QDialog>
@@ -29,7 +31,7 @@ class QLineEdit;
 class QDoubleSpinBox;
 class ParameterDialog : public QDialog {
 public:
-    ParameterDialog(const ParameterBoundary& boundary, double value, QWidget* parent = 0);
+    ParameterDialog(const ParameterBoundary& boundary, double value, double SSE, const QString& parameter, QWidget* parent = 0);
 
     void setUi();
     virtual void accept() override;
@@ -39,12 +41,18 @@ public:
 private:
     void showParameter();
     void hideParameter();
-    QDialogButtonBox* m_buttonbox;
+    void Recalculate();
+    void Adopt();
 
+    QDialogButtonBox* m_buttonbox;
+    QPushButton* m_adopt_and_accept;
     QDoubleSpinBox *m_value_edit, *m_maximum_edit, *m_minimum_edit, *m_lower_barrier_beta, *m_lower_barrier_wall, *m_upper_barrier_beta, *m_upper_barrier_wall;
     QCheckBox *m_lower_limit, *m_upper_limit;
-
+    QLabel *m_general_information, *m_left_boundary, *m_right_boundary;
+    ChartView* m_chart;
+    LineSeries *m_lowerbound, *m_upperbound, *m_combined, *m_sse_series, *m_current_value, *m_current_penalty;
     ParameterBoundary m_boundary;
-    double m_value;
+    QString m_parameter;
+    double m_value = 0, m_SSE = 0;
     int m_focus = 0;
 };